requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.1
django-filter>=24.2
requests>=2.31
django-import-export>=4.0
drf-spectacular>=0.28
gunicorn>=22.0
whitenoise
requests>=2.31.0